using the Z correction of 1/5 res in Z to flatten

+++ globals.py
@@ -12,4 +12,4 @@ globals.py - global defines that don't change.
 
 import torch
 
-DTYPE = torch.float16
+DTYPE = torch.float32
